requirements.txt
SQLAlchemy
alembic
psycopg2-binary
pandas
tqdm
# Environment variable management
python-dotenv
python-binance
cryptography